+++ crates/sage-cloudpath/src/lib.rs
@@ -9,6 +9,9 @@ use tokio::io::{AsyncBufRead, AsyncRead, AsyncWriteExt, BufReader};
 pub mod mzml;
 pub mod util;
 
+#[cfg(feature = "parquet")]
+pub mod parquet;
+
 static S3_CLIENT: once_cell::sync::OnceCell<aws_sdk_s3::Client> = once_cell::sync::OnceCell::new();
 
 async fn s3_client() -> &'static aws_sdk_s3::Client {